Remove Unused Code
